requirements.txt
certifi==2021.5.30
gevent==21.1.2
greenlet==1.1.0
kafka-python==2.0.2
pymongo==3.12.0
python-dotenv==0.19.0
websocket-client==1.1.0
zope.event==4.5.0
zope.interface==5.4.0
